Fix merge issues and move GetVersion into generic conf
