+++ cmd/genericconf/getversion17.go
@@ -1,6 +1,6 @@
 //go:build !go1.18
 
-package conf
+package genericconf
 
 func GetVersion() (string, string) {
 	return "development", "development"
